Change FILE* to fstream

+++ src/BufferPipeline.h
@@ -5,10 +5,10 @@ inline const int LENGTH_OF_BUFFER = 44100*60*4; // 44100 - average frequency, 4 
 class BufferPipeline{
 public:
     short* buffer;
-    unsigned int pos;
+    unsigned long long pos;
     unsigned int currSec;
     unsigned int frequency;
-    unsigned int endPos;
+    unsigned long long endPos;
     BufferPipeline();
     ~BufferPipeline();
 };
